Move main project into its own subfolder
