+++ endiffo.sln
@@ -3,23 +3,35 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27703.2026
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "endiffo", "endiffo.csproj", "{6627C567-0ABE-4EC2-871F-B2F780967892}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "endiffo", "endiffo\endiffo.csproj", "{1D8495B3-2531-4228-A664-AC9D8D4B938A}"
 EndProject
 Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6627C567-0ABE-4EC2-871F-B2F780967892}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6627C567-0ABE-4EC2-871F-B2F780967892}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6627C567-0ABE-4EC2-871F-B2F780967892}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6627C567-0ABE-4EC2-871F-B2F780967892}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {435107C9-3C05-45A4-8F0C-07819F8D6110}
+	EndGlobalSection
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Debug|x64.Build.0 = Debug|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Debug|x86.Build.0 = Debug|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Release|x64.ActiveCfg = Release|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Release|x64.Build.0 = Release|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Release|x86.ActiveCfg = Release|Any CPU
+		{1D8495B3-2531-4228-A664-AC9D8D4B938A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
