+++ endiffo/Constants.cs
