+++ endiffo/Registry.cs
