+++ endiffo/Search/Collection.cs
