+++ endiffo/Search/ISearch.cs
