+++ endiffo/Search/RegistryScan.cs
