+++ endiffo/Search/SimpleScan.cs
