+++ endiffo/Utility.cs
