+++ endiffo/Worker/Writer.cs
